Counter component creation
