+++ src/components/home/NewItems.jsx
@@ -5,6 +5,7 @@ import axios from 'axios';
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
+import Counter from "../UI/Counter";
 
 
 const NewItems = () => {
@@ -58,17 +59,7 @@ const NewItems = () => {
   }, []);
 
 
-  // Countdown formatting function in seconds
-  const formatCountdown = (endTimeMilliseconds) => {
-    const currentTime = Date.now()
-    const timeRemaining = Math.max(0, endTimeMilliseconds - currentTime);
-    const timeInSeconds = Math.floor(timeRemaining / 1000);
-    
-    const hours = Math.floor(timeInSeconds / 3600);
-    const minutes = Math.floor((timeInSeconds % 3600) / 60);
-    const seconds = timeInSeconds % 60;
-    return `${hours}h ${minutes}m ${seconds}s`;
-  };
+
 
 
   // Loading skeleton component
@@ -128,9 +119,12 @@ const NewItems = () => {
                             <i className="fa fa-check"></i>
                           </Link>
                         </div>
-                        <div className="de_countdown"> 
-                          {item.expiryDate ? formatCountdown(item.expiryDate) : "Expired"} 
-                          </div>
+
+
+                          <Counter />
+
+
+
                         <div className="nft__item_wrap">
                           <div className="nft__item_extra">
                             <div className="nft__item_buttons">
